Undo the previous change in 'integer*2 ref1, ref2' -- wait for HDF group decision on the number type for the ref's in xxrref's and xxwref's.

+++ hdf/test/tr8f.f
@@ -34,7 +34,7 @@ C
 
       integer x, y, ret, num_images, number_failed
       integer d1, d2, ispal, FALSE, TRUE
-      integer*2 ref1, ref2
+      integer ref1, ref2
       integer  ref3
 
       DFTAG_RLE = 11
